Update README.md
Changes in communication with CHINT allow you to use the original MODBUS libraries without any modifications.
Changing the address from 0x3F80 to 0x3B11 ensures compatibility with older and newer firmware for SUN2000.
Since the ESP32 power supply is 3. 3 V and we use it to power RS485 interfaces, remove the resistor between A and B.  If RS485 interfaces are supplied with higher voltage, these resistors do not need to be removed. This is due to the current specification of the RS485 standard.